Always call processReplyBody() after adaptOrFinalizeReply().
The recent fix for "store_status == STORE_PENDING" added a check
with return from HttpStateData::processReply() before
processReplyBody() call.  But processReplyBody() does proper
server state cleanup.  We could end up with server state that
would not be cleaned (until connection is closed).  That could
also introduce other issues.  The patch moves the isAccepting()
check from processReply() to processReplyBody().  We do not write
to an entry that is not accepting, but allow processReplyBody()
to do proper cleanup.