author: Measurement Factory Bug 3118: ecap_enable on forces icap_enable on
We were updating [Icap|Ecap]::TheConfig even when [icap|ecap]_enable was false,
which may lead to service activation for Icap or Ecap services that should be
disabled. The patch removes such services from service groups before they are 
activated.

The patch also warns the user when an adaptation group loses some but not all
of its services due to the new group cleanup code.