Warn about Rock cache_dir disk space waste.
In some configurations Rock cache_dir may hit the maximum entry limit.
In that case a significant disk space may be wasted.  Before recent
SMP-related changes, there was code to warn the user about such
situations.  The patch resurrects that code, adjusting it as needed to
match the current realities.