Simplify Ssl::ServerPeeker and renamed it to Ssl::ServerBump
Currently, Ssl::ServerPeeker is a job that does virtually nothing and leaks on 
SSL connection establishment errors. This patch convert this class convert this
class into a non-job class which ConnStateData use to maintain SSL
certificate peeking state.
The Ssl::ServerPeeker class renamed  to Ssl::ServerBump.
ConnStateData::bumpErrorEntry and ConnStateData::bumpSslErrorNoList and 
ConnStateData::bumpServerCert now are members to the this class.