autoconf upgrade: basic_auth helpers (and Bug 3793)
Convert the basic_auth bundled helpers from using bash to m4 scripts
when testgn which helpers are able to be built.

This conversion allows us to leverage existing information found by
./configure in the helper test scritps, and to de-duplicate and simplify
many of the testing checks.

This required shuffling the SASL checks into a custom macro, which is
altered from a halting ERROR when SASL is absent to a softer WARNING,
but prevents the SASL helpers building - which the old checks did not
guarantee. Resolving Bug 3793.

The nature of m4 macros adds two small long term issues. Both
well-known issues with m4_include() expansion:
 * the macro is expanded in-place without scoping separation.
 So any brokenness in the child scripts will be a little harder to
 identify and may result in breakage of modues other than the one whose
 child script was the cause.

 * the macro takes literal arguments without variable expansion.
 So a "full" path relative to the main configure.ac script is required
 in every include. Meaning the modules.m4 file cannot simply
 sub-include using the $helper script variable, but must be a full
 expansion of the modules set.