Added Expect:100-continue to FTP STOR wrappers for adaptation to avoid stalling FTP clients that expect an FTP 150 (OK to send data) response from the FTP server before actually uploading content.
If an adaptation service starts waiting for the PUT request body before
starting to produce an adaptation response (allowing the transaction to reach
the FTP server), the FTP transaction will get stuck. It is not clear why this
has not been a [known] issue with real HTTP requests containing
Expect:100-continue.