To minimize TCP races, delay complaining about missing passive data connection to Squid until we talk to the server.
After PASV, it is possible for the client TCP handshake to reach Squid _after_
a LIST or RETR command is parsed, even if the client properly initiates the
connection before writing the command. We no longer immediately respond with
425 "Data connection is not established" in such cases, but re-check the data
connection availability after we talk to the server.

Also polished/fixed a few comments.