Cleanup: rewrite Http::One::Parser::getHeaderField() using Tokenizer
Fixes performance regression in SBuf usage.

Also, fixes one nasty bug where it would return the value of a line
containign the named header despite obs-fold and quoting. This was
particualarly bad as this method is used primarily to retrieve Host:
header before full aprsing of the mime block takes place.

The bug causes "foo.invalid" to be detected as hostname in:

  GET / HTTP/1.1
  Foo: bar="\r  Host: fake.invalid\n  "
  Host: example.com