Peek and Splice: %ssl::<cert_subject and %ssl::<cert_issuer formating codes to e xternal_acl
This patch investigates the %ssl::<cert_subject and %ssl::<cert_issuer 
formating codes to external_acl helpers.
 * The %ssl::<cert_subject formating code prints the server certificate DN
 * The %ssl::<cert_issuer formating code prints the server certificate issuer DN

Both formating codes are available after the ssl bumped connection is 
established.
When Peek and Splice mode is selected these formating codes are available on
peek or stare mode, after the step2 is comleted and server certificate is 
received.