Convert HttpStateData (aka. Http::Client) response buffer to use SBuf
This replaces the MemBuf used to receive server responses with an SBuf in
preparation for use by an Http1::ResponseParser.

As part of this convert the I/O read processing to Comm::Read / ReadNow
API with read(2) operation happening in the read handler instead of
buried in the comm code. A side effect of this is detaching the read
operation from StoreEntry::delayAwareRead().

Two temporary performance regressions are necessarily introduced with
these changes:
* The worst is a data copy from SBuf input buffer to a temporary MemBuf
  so that the chunked body decoder can process the content. This will
  only be able to be removed when teh chunked decoder is itself converted
  to process an SBuf.

* Also, until the new Http1::ResponseParser is implemented we need to
  copy the response header contents into a temporary MemBuf for the old
  HttpReply::parse() method to process.