Fix delay pool DeferredRead in HttpStateData
The deferred read action may have been delayed so long the read(2) is no
longer possible. It needs to restart from watching the socket instead of
just from performing read(2). This re-adds part of the latency when
delaying traffic but we retain the bytesWanted() speed up so not a big
loss.